pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quantframe"
version = "0.1.0"
description = "A comprehensive quantitative trading framework"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["trading", "quantitative", "finance", "backtesting"]
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "yfinance>=0.2.0",
    "ta>=0.10.0",  # Technical analysis
    "pyfolio>=0.9.2",  # Performance analytics
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "fastapi>=0.68.0",  # For API endpoints
    "pydantic>=1.8.0",  # Data validation
    "joblib>=1.0.0",  # Parallel processing
    "sqlalchemy>=1.4.0",  # Database operations
    "pytest>=6.0.0",
    "black>=21.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
    "python-binance>=1.0.19",  # Binance API
    "ccxt>=4.0.0",  # Unified crypto exchange API
    "aiohttp>=3.8.0",  # Async HTTP client
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
    "pytest-cov>=2.12.0",
    "pre-commit>=2.15.0",
]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3
